v1.2
Added a whole new version of the script: a pygubu-based GUI launcher. Due to some limitations (which I may find a workaround for in the future), you can only launch 1 account at a time with this version of the launcher, so for now, the old command prompt-based launcher is also being upkept. As you can see, the GUI Launcher is largely based on the old command prompt launcher's code. So, while I was making this, I went ahead and cleaned up the code in the old launcher as well, removing a couple of deprecated functions. Also, one last feature, both launchers should automatically install all the libraries if you don't have them installed, rather than making you do it yourself.